feat(tokens): add tokens sort by fiat amount
